Refactoring + PageRequest struct
