+++ massa-api/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-paginate = "1.1.11"
 jsonrpsee = { version = "0.16.2", features = ["server", "macros"] }
 async-trait = "0.1.58"
 serde = { version = "1.0", features = ["derive"] }
